Fix example code for constructing ASTs.
Fixed the second example so it produces the same AST as the first. Also, I figured out that the variable `ast` was just the same as `tree`.

Didn't do this, but the second example could also be formatted like the first to more clearly show the structure of the calls.